+++ Editor/AutoRunner/AutoRunnerResultsRenderer/ResultsRenderer.cs
@@ -3,9 +3,7 @@ using SaintsField.Editor.Playa.Renderer;
 using UnityEditor;
 using System.Collections.Generic;
 using System.Linq;
-using System;
 using SaintsField.Editor.Playa.Renderer.ShowInInspectorFieldFakeRenderer;
-using SaintsField.Editor.Utils;
 using Object = UnityEngine.Object;
 
 namespace SaintsField.Editor.AutoRunner.AutoRunnerResultsRenderer
@@ -18,27 +16,6 @@ namespace SaintsField.Editor.AutoRunner.AutoRunnerResultsRenderer
         {
             RenderField = true;
             _autoRunner = (AutoRunnerWindowBase) serializedObject.targetObject;
-        }
-
-        private struct MainTarget : IEquatable<MainTarget>
-        {
-            public string MainTargetString;
-            public bool MainTargetIsAssetPath;
-
-            public bool Equals(MainTarget other)
-            {
-                return MainTargetString == other.MainTargetString && MainTargetIsAssetPath == other.MainTargetIsAssetPath;
-            }
-
-            public override bool Equals(object obj)
-            {
-                return obj is MainTarget other && Equals(other);
-            }
-
-            public override int GetHashCode()
-            {
-                return Util.CombineHashCode(MainTargetString, MainTargetIsAssetPath);
-            }
         }
 
         private struct AutoRunnerResultInfo