+++ Runtime/Playa/OrderedAttribute.cs
@@ -4,6 +4,7 @@ using System.Runtime.CompilerServices;
 
 namespace SaintsField.Playa
 {
+    [Obsolete("SaintsField now can detect order correctly, this attribute is no longer required")]
     [Conditional("UNITY_EDITOR")]
     [AttributeUsage(AttributeTargets.Field | AttributeTargets.Method | AttributeTargets.Property)]
     public class OrderedAttribute: Attribute, IPlayaAttribute